requirements.txt
# Core Streamlit and LangChain dependencies
streamlit==1.46.0
langchain==0.1.12
langchain-community==0.0.29
langchain-core==0.1.36
langchain-google-genai==0.0.11
langchain-text-splitters==0.0.2
langchain-huggingface==0.1.2

#LlamaIndex and LlamaParse dependencies
llama-index==0.12.42
llama-index-llms-google-genai==0.2.1
llama-index-embeddings-google-genai==0.2.0
llama-index-vector-stores-opensearch==0.5.6
llama-parse==0.6.34

# OpenSearch vector database
opensearch-py==2.7.0
langchain-opensearch==0.0.2

# Google AI/Gemini dependencies
google-generativeai==0.4.1
google-ai-generativelanguage==0.4.0
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.66.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0

# Microsoft Graph API
msal==1.27.0
msgraph-core==0.2.2
O365==2.0.26

# PDF processing
PyMuPDF==1.22.5
pdfminer.six==20231228
pdfplumber==0.11.5
pypdf==5.3.0

# Environment and configuration
python-dotenv==1.0.1

# Core data science packages
numpy==1.26.4
pandas==2.2.1
scipy==1.15.1

# HTTP and networking
requests==2.32.3
httpx==0.28.1
urllib3==1.26.20

# Pydantic for data validation
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2

# Async and concurrency
anyio==4.8.0
aiohttp==3.11.12
aiosignal==1.3.2

# Utility packages
typing_extensions==4.12.2
packaging==23.2
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
click==8.1.8
attrs==25.1.0

# Optional: Keep FAISS for fallback (if needed)
faiss-cpu==1.10.0

# Streamlit dependencies
altair==5.5.0
blinker==1.9.0
cachetools==5.5.1
gitdb==4.0.12
GitPython==3.1.44
importlib_metadata==8.5.0
Jinja2==3.1.5
jsonschema==4.23.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
pillow==11.2.1
protobuf==4.25.3
pyarrow==19.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
rich==13.9.4
tenacity==8.2.3
tornado==6.4.2
tzdata==2025.1
validators==0.34.0
watchdog==6.0.0

# Google Sheets API dependencies
openpyxl==3.1.2
gspread==6.0.2
oauth2client==4.1.3

# Additional dependencies
spacy>=3.7.0
numpy>=1.24.0